Move "fs" utils to own module.
