+++ src/client/interpreter/locators/helpers.ts
@@ -2,7 +2,8 @@ import { inject, injectable } from 'inversify';
 import * as path from 'path';
 import { getArchitectureDislayName } from '../../common/platform/registry';
 import { IFileSystem, IPlatformService } from '../../common/platform/types';
-import { fsReaddirAsync, IS_WINDOWS } from '../../common/util';
+import { IS_WINDOWS } from '../../common/util';
+import { fsReaddirAsync } from '../../common/utils/fs';
 import { IServiceContainer } from '../../ioc/types';
 import { IInterpreterHelper, IInterpreterLocatorHelper, InterpreterType, PythonInterpreter } from '../contracts';
